refactor(types): mypy-Sweep Tranche 1 — risk/ + paper/ auf 0 Fehler, CI-Gate-Scope erweitert
GESAMTBEWERTUNG par.8: erste Tranche des mypy-Sweeps. risk/ 41->0 und
paper/ 4->0 Fehler (44 Fixes), STRENG verhaltensneutral: nur Annotations,
explizite float()/cast()-Wrapper (Cast-Positionen verifiziert, kein
float(x/y)-Trap), None-Guards die den bestehenden Vertrag explizit machen
(Barra-Asserts mit Message, pre-commit-konform; fit() setzt unconditional;
Modul hat 0 Produktions-Caller), unused-ignores raus, gezielte
import-ignores. Turnover-Renames = reine mypy-Redefinitions-Fixes.

ZWEI potenzielle ECHTE Bugs bewusst NICHT gefixt, als FIXME(mypy-sweep)
markiert (Fix waere Verhaltensaenderung, braucht eigenen Auftrag):
- paper/strategy_adapters.py: params.get("ma_fast") kann None die
  Funktions-Defaults ueberschreiben -> Downstream-Crash statt Default.
- paper/paper_track.py: TradingContext(output_dir=None) gegen
  Path-Deklaration; richtiger Fix liegt in pipeline/ (Schutzzone).
  Stage-1-Zusatzfund: _tc_execution-Heartbeat dereferenziert output_dir
  OHNE write_outputs-Guard (pre-existing, Follow-up).
- risk_metrics: NOTE dass ml/evt_models ARCHIVIERT ist (d2c3d093) —
  EVT-Pfad = toter Code mit ehrlichem unavailable-Fallback.

backend-ci mypy-Gate: Scope +risk +paper. Delta-Beweis (Stage-1 UND
Stage-2 unabhaengig reproduziert): 5-Pfad und 7-Pfad melden identische
72 followed-import-Fehler (alle in accounting/ops/pipeline/qa, 0 in
risk/paper), 0 neue -> Flip delta-neutral. OFFENES Follow-up (F-senior-2):
lokal zeigt schon das ALTE Gate 72 followed-Fehler, CI ist gruen —
Env-Divergenz; dieser Push liefert den CI-Beweis fuer den Flip.

Review-Kette: Stage-1 risk-execution-reviewer (Verhaltensneutralitaet
BESTAETIGT, 23 Tests selbst ausgefuehrt) + Stage-2 senior-code-reviewer
(Flip reproduziert, commit-ready). Deny-Lift dokumentiert, settings.json
byte-identisch restauriert.

Tests: Sweep 1200 passed/4 skipped + 30 DD/Kill-Switch + Stage-1 23 +
final 167 Zusatztests 0 failed. CI-Bestaetigung: dieser Push.

+++ src/assembled_core/risk/garch_vol.py
@@ -27,7 +27,7 @@ logger = logging.getLogger(__name__)
 
 _ARCH_AVAILABLE: bool
 try:
-    from arch import arch_model as _arch_model  # type: ignore  # noqa: F401
+    from arch import arch_model as _arch_model  # noqa: F401
 
     _ARCH_AVAILABLE = True
 except ImportError:
@@ -85,7 +85,7 @@ def forecast_vol(
 
 def _garch_forecast(r: pd.Series, horizon: int, annualize_factor: float) -> float:
     # arch_model expects returns in percent (it rescales internally)
-    from arch import arch_model  # type: ignore
+    from arch import arch_model
 
     model = arch_model(
         r * 100,
@@ -102,14 +102,14 @@ def _garch_forecast(r: pd.Series, horizon: int, annualize_factor: float) -> floa
     mean_var_pct2 = float(fc.variance.iloc[-1].mean())
     # Convert back: std in fraction = sqrt(var_pct2) / 100
     sigma_per_bar = np.sqrt(mean_var_pct2) / 100.0
-    return sigma_per_bar * np.sqrt(annualize_factor)
+    return float(sigma_per_bar * np.sqrt(annualize_factor))
 
 
 def _fallback_vol(r: pd.Series, window: int, annualize_factor: float) -> float:
     tail = r.tail(window)
     if len(tail) < 2:
         return float("nan")
-    return float(tail.std(ddof=1)) * np.sqrt(annualize_factor)
+    return float(float(tail.std(ddof=1)) * np.sqrt(annualize_factor))
 
 
 # ---------------------------------------------------------------------------
